pyproject.toml
[project]
name = "hass-mcp"
version = "0.2.0"
description = "Home Assistant Model Context Protocol (MCP) server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.4.1",
    "homeassistant-api>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
